Moved heartbeat config to support package
